Added temp fix to prevent app crashing when Join CIS fails

+++ platform-infrastructure/client-frwk/SocietiesAndroidApp/assets/www/javascript/pagespecific/communities_list.js
@@ -183,13 +183,14 @@ var	SocietiesCISListService = {
 		
 		//SEND REQUEST
 		var activity = mActivities[pos];
-		//if (window.confirm("Remove this entry?")) {
-		jConfirm("Delete this entry?", 'Delete Activity', function(answer) {
+		if (window.confirm("Remove this entry?")) {
+		//jConfirm("Delete this entry?", 'Delete Activity', function(answer) {
 			if (answer) {
 				$('#li' + mActivities[pos].published).append("Removing...");
 				window.plugins.SocietiesLocalCISManager.deleteActivity(cisId, activity, success, failure);
 			}
-		});
+		}
+		//);
 	},
 
 	/*
@@ -329,12 +330,13 @@ var	SocietiesCISListService = {
 		}
 		
 		//SEND REQUEST
-		//if (window.confirm("Send friend request to " + name + "?")) {
-		jConfirm("Send friend request to " + name + "?", 'Friend Request', function(answer) {
+		if (window.confirm("Send friend request to " + name + "?")) {
+		//jConfirm("Send friend request to " + name + "?", 'Friend Request', function(answer) {
 		     if (answer){
 		    	 window.plugins.SocietiesLocalCSSManager.sendFriendRequest(css_id, success, failure);
 		     }
-		});
+		}
+		//);
 	},
 	
 	/**
@@ -388,6 +390,8 @@ var	SocietiesCISListService = {
 				'</li>';
 			$('ul#cis_shared_apps').append(tableEntry);
 			$('ul#cis_shared_apps').listview('refresh');
+			//REFRESH ACTIVITY FEED
+			SocietiesCISListService.showCISActivities(mCis_id, false);
 		}
 		
 		function failure(data) {
@@ -401,15 +405,16 @@ var	SocietiesCISListService = {
 		if (servicePos != "0000") { //"Select a Service"
 			mCis_id = $('input#cis_id').val();
 			serviceObj = mMyServices[servicePos];
-			//if (confirm("Share service: " + serviceName + " to this community?")) {
-			jConfirm("Share " + serviceName + " to this community?", 'Share Service', function(answer) {
+			if (confirm("Share " + serviceName + " to this community?")) {
+			//jConfirm("Share " + serviceName + " to this community?", 'Share Service', function(answer) {
 			     if (answer){
 					window.plugins.ServiceManagementService.shareMyService(mCis_id, serviceObj, success, failure);
 					//ADD SERVICE TO LIST OF SHARED SERVICES
 					mCisServices.push(serviceObj);
 			     } else
 				$('select#selShareService').attr('selectedIndex', 0);
-			});
+			}
+			//);
 		}
 		$('select#selShareService').selectmenu('refresh');
 	},
